# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "TowTruckMove REST API",
    "description": ""
  },
  "host": "towtruckmove-api.onrender.com",
  "basePath": "/",
...
  "schemes": ["https"],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/user/create": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Email": {
                  "example": "any"
                },
                "CompanyName": {
                  "example": "any"
                },
                "ContactEmail": {
                  "example": "any"
                },
                "ContactPhone": {
                  "example": "any"
                },
                "CompanyAddress": {
                  "example": "any"
                },
                "Region": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "RoleType": {
                  "example": "any"
                },
                "Specilaization": {
                  "example": "any"
                },
                "Password\r": {
                  "example": "any"
                },
                "Password": {
                  "example": "any"
                },
                "FullName": {
                  "example": "any"
                },
                "FirstName": {
                  "example": "any"
                },
                "LastName": {
                  "example": "any"
                },
                "Phone": {
                  "example": "any"
                },
                "Address": {
                  "example": "any"
                },
                "AcceptTerms": {
                  "example": "any"
                },
                "PaymentMethod": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findOne/{userId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAll": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllBySearch/{name}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllUsersByDate/{startDate}/{toDate}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/update/{userId}": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "\r/api/user/updateFile": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                },
                "FileType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/user/findRoles": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findUserRoles": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateRole/{roleId}": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateUserRole/{userId}": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/deleteRole/{roleId}": {
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/subscribe": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                },
                "SubscriptionId": {
                  "example": "any"
                },
                "SubscriptionName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/user/upgradeUserSubscription": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserSubscriptionId": {
                  "example": "any"
                },
                "UserId": {
                  "example": "any"
                },
                "SubscriptionId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateUserSubscription/{userSubscriptionId}": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "userSubscriptionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserSubscriptionId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findUserSubscription/{userId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllUserSubscriptions/{subscriptionId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllUserSubscriptionsByDate/{startDate}/{toDate}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "\r/api/user/findAllUserSubscriptionsByStartDate/{startDate}/{toDate}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "\r/api/user/findAllUserSubscriptionsByEndDate/{startDate}/{toDate}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/createCompany": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CompanyName": {
                  "example": "any"
                },
                "ContactEmail": {
                  "example": "any"
                },
                "ContactPhone": {
                  "example": "any"
                },
                "Address": {
                  "example": "any"
                },
                "Country": {
                  "example": "any"
                },
                "CompanyType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateCompany/{companyId}": {
      "put": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CompanyId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "\r/api/user/company/uploadCompanyDoc": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "UserId": {
                  "example": "any"
                },
                "FileType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findCompany/{companyId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllCompanys/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "companyType",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/findAllCompanysByDate/{startDate}/{toDate}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/deleteCompany/{companyId}": {
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}